+++ src/main/java/fr/insalyon/algorithm/TemplateTSP.java
@@ -7,7 +7,7 @@ import java.util.Iterator;
 public abstract class TemplateTSP implements TSP {
 	private Integer[] bestSol;
 	protected Graph g;
-	private int bestSolCost;
+	private float bestSolCost;
 	private int timeLimit;
 	private long startTime;
 	
@@ -25,13 +25,13 @@ public abstract class TemplateTSP implements TSP {
 		branchAndBound(0, unvisited, visited, 0);
 	}
 	
-	public Integer getSolution(int i){
+	public int getSolution(int i){
 		if (g != null && i>=0 && i<g.getNbVertices())
 			return bestSol[i];
 		return -1;
 	}
 	
-	public int getSolutionCost(){
+	public float getSolutionCost(){
 		if (g != null)
 			return bestSolCost;
 		return -1;
@@ -63,9 +63,9 @@ public abstract class TemplateTSP implements TSP {
 	 * @param currentCost the cost of the path corresponding to <code>visited</code>
 	 */	
 	private void branchAndBound(int currentVertex, Collection<Integer> unvisited, 
-			Collection<Integer> visited, int currentCost){
+			Collection<Integer> visited, float currentCost){
 		if (System.currentTimeMillis() - startTime > timeLimit) return;
-	    if (unvisited.size() == 0){ 
+		if (unvisited.size() == 0){
 	    	if (g.isArc(currentVertex,0)){ 
 	    		if (currentCost+g.getCost(currentVertex,0) < bestSolCost){ 
 	    			visited.toArray(bestSol);
